+++ contracts/books/boh/ShareholdersAgreement.sol
@@ -305,14 +305,14 @@ contract ShareholdersAgreement is
 
     // ==== VotingRules ====
 
-    function votingRules(uint8 typeOfVote) external view returns (bytes32) {
+    function votingRules(uint16 typeOfVote) external view returns (bytes32) {
         require(typeOfVote != 0, "SA.votingRules: zero typeOfVote");
         require(typeOfVote < 21, "SA.votingRules: typeOfVote over flow");
 
         return _rules[typeOfVote];
     }
 
-    function additionalVetoholdersOfVR(uint8 typeOfVote)
+    function additionalVetoholdersOfVR(uint16 typeOfVote)
         external
         view
         returns (uint40[] memory)
@@ -325,21 +325,21 @@ contract ShareholdersAgreement is
 
     // ==== FirstRefusal Rule ====
 
-    function isSubjectToFR(uint8 typeOfDeal) external view returns (bool) {
+    function isSubjectToFR(uint16 typeOfDeal) external view returns (bool) {
         require(typeOfDeal != 0, "SHA.isSubjectToFR: typeOfDeal overflow");
         require(typeOfDeal < 4, "SHA.isSubjectToFR: typeOfDeal overflow");
 
         return _seqOfRules.contains(20 + typeOfDeal);
     }
 
-    function ruleOfFR(uint8 typeOfDeal) external view returns (bytes32) {
+    function ruleOfFR(uint16 typeOfDeal) external view returns (bytes32) {
         require(typeOfDeal != 0, "SHA.ruleOfFR: typeOfDeal overflow");
         require(typeOfDeal < 4, "SHA.ruleOfFR: typeOfDeal overflow");
 
         return _rules[20 + typeOfDeal];
     }
 
-    function isRightholderOfFR(uint8 typeOfDeal, uint40 acct)
+    function isRightholderOfFR(uint16 typeOfDeal, uint40 acct)
         external
         view
         returns (bool)
@@ -353,7 +353,7 @@ contract ShareholdersAgreement is
         else return _rightholders[typeOfDeal + 20].contains(acct);
     }
 
-    function rightholdersOfFR(uint8 typeOfDeal)
+    function rightholdersOfFR(uint16 typeOfDeal)
         external
         view
         returns (uint40[] memory)
